test: achieve 100% real-code coverage through state-machine integration and Win32 lifecycle tests
